made http request handler an independent class, removed global arg, responded to a code review comment
